rename reducers/events -> reducers/data; rename state.events -> state.data;
