+++ src/selectors/header-banner.js
@@ -2,10 +2,8 @@
 import type { State } from "../reducers";
 import type { HeaderBanner } from "../data/header-banner";
 
-const getEventsState = (state: State) => state.events;
-
 /* eslint-disable import/prefer-default-export */
 export const selectHeaderBanners = (state: State): HeaderBanner[] =>
-  ((getEventsState(state).entries.filter(
+  ((state.data.entries.filter(
     entry => entry.sys.contentType.sys.id === "headerBanner"
   ): any[]): HeaderBanner[]);
