+++ src/selectors/header-banner.test.js
@@ -3,7 +3,7 @@ import { selectHeaderBanners } from "./header-banner";
 describe("selectHeaderBanners", () => {
   it("selects property", () => {
     const state = {
-      events: {
+      data: {
         entries: [
           { sys: { contentType: { sys: { id: "event" } } } },
           { sys: { contentType: { sys: { id: "headerBanner" } } } }
@@ -14,6 +14,6 @@ describe("selectHeaderBanners", () => {
     const selected = selectHeaderBanners(state);
 
     expect(selected.length).toBe(1);
-    expect(selected[0]).toBe(state.events.entries[1]);
+    expect(selected[0]).toBe(state.data.entries[1]);
   });
 });
