+++ src/selectors/sponsors.js
@@ -2,10 +2,10 @@
 import type { State } from "../reducers";
 import type { Sponsor } from "../data/sponsor";
 
-const getEventsState = (state: State) => state.events;
+const getDataState = (state: State) => state.data;
 
 /* eslint-disable import/prefer-default-export */
 export const selectSponsors = (state: State): Sponsor[] =>
-  ((getEventsState(state).entries.filter(
+  ((getDataState(state).entries.filter(
     entry => entry.sys.contentType.sys.id === "sponsor"
   ): any[]): Sponsor[]);
