+++ src/selectors/sponsors.test.js
@@ -3,7 +3,7 @@ import { selectSponsors } from "./sponsors";
 describe("selectSponsors", () => {
   it("selects property", () => {
     const state = {
-      events: {
+      data: {
         entries: [
           { sys: { contentType: { sys: { id: "event" } } } },
           { sys: { contentType: { sys: { id: "sponsor" } } } }
@@ -14,6 +14,6 @@ describe("selectSponsors", () => {
     const selected = selectSponsors(state);
 
     expect(selected.length).toBe(1);
-    expect(selected[0]).toBe(state.events.entries[1]);
+    expect(selected[0]).toBe(state.data.entries[1]);
   });
 });
